@@ -21,13 +21,13 @@
     "cookie": "^0.5.0",
     "cookies-next": "^2.1.1",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
+    "eslint-config-next": "^13.4.2",
     "inversify": "^6.0.1",
-    "next": "13.3.0",
+    "next": "^13.4.2",
     "next-auth": "^4.22.1",
     "next-redux-wrapper": "^8.1.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.9",
     "react-modal": "^3.16.1",
     "react-redux": "^8.0.5",
