@@ -2,7 +2,7 @@ import { Action, ThunkAction, configureStore } from "@reduxjs/toolkit";
 import { userSlice } from "../reducer/UserReducer";
 import { createWrapper } from "next-redux-wrapper";
 import { messagesSlice } from "../reducer/MessagesReducer";
-const makeStore = () =>
+export const store = () =>
   configureStore({
     reducer: {
       [userSlice.name]: userSlice.reducer,
@@ -11,7 +11,7 @@ const makeStore = () =>
     devTools: true,
   });
 
-export type AppStore = ReturnType<typeof makeStore>;
+export type AppStore = ReturnType<typeof store>;
 export type AppState = ReturnType<AppStore["getState"]>;
 export type AppTunk<ReturnType = void> = ThunkAction<
   ReturnType,
@@ -20,4 +20,4 @@ export type AppTunk<ReturnType = void> = ThunkAction<
   Action
 >;
 
-export const wrapper = createWrapper<AppStore>(makeStore);
+export const wrapper = createWrapper<AppStore>(store);
